==> src/20211028_freeze/6_plots_transferability.ipynb <==
# PRS map - summary plots

Yosuke Tanigawa


## library, functions, and constants

suppressWarnings(suppressPackageStartupMessages({
    library(tidyverse)
    library(data.table)
}))


source('paths.sh')
devtools::load_all(cud4_d)
devtools::load_all(dirname(dirname(snpnet_helper)))
source(snpnet_helper)


traits_w_metrics_f %>%
fread() %>%
rename_with(
    function(x){str_replace(x, '#', '')}, starts_with("#")
) -> traits_w_metrics_df


eval_fullwDelta_f %>%
fread() %>%
rename_with(
    function(x){str_replace(x, '#', '')}, starts_with("#")
) -> eval_long_df


get_split_order_df <- function(){
    data.frame(
        split = c('non_british_white', 's_asian', 'e_asian', 'african'),
        split_plot = c('Non-British white', 'South Asian', 'East Asian', 'African'),
        split_order=1:4, stringsAsFactors=F
    )
}


get_split_color_dict <- function(){
    setNames(
        c(cud4_colors()['gray'], cud4_colors()['purple'], cud4_colors()['sky_blue'], cud4_colors()['orange'] ),
        c('Non-British white', 'South Asian', 'East Asian', 'African')
    )
}


transferability_eval_func <- function(ancestry_group_metric, WB_test_set_metric){
    ancestry_group_metric - WB_test_set_metric
}


eval_long_df %>%
filter(
    split %in% c('test', 'non_british_white', 'african', 's_asian', 'e_asian')
) %>% 
select(all_of(c("trait", "split", "metric", "model", "value"))) %>% 
pivot_wider(
    names_from = c("split"),
    values_from = c("value")
) %>% 
mutate(across(
    c('non_british_white', 'african', 's_asian', 'e_asian'),
    function(x){ transferability_eval_func(x, test) }
)) %>%
pivot_longer(
    all_of(c('non_british_white', 'african', 's_asian', 'e_asian')),
    names_to = "split",
    values_to = "delta_vs_WBtest"
) %>% 
drop_na(delta_vs_WBtest) %>%
rename(
    'delta_WBtest'='test'
) %>%
left_join(
    traits_w_metrics_df %>% 
    select(all_of(c(
        'trait', 'trait_name', 'trait_category',
        'trait_category_plot', 'trait_label',
        'family', 'WBtest_is_significant'
    ))) %>% unique(),
    by = "trait"
) %>%
left_join(
    get_split_order_df(),
    by = "split"
) -> transferability_df


transferability_df %>%
filter(
    WBtest_is_significant,
    metric %in% c("r2", "NagelkerkeR2"),
    model == "delta"
) %>%
count(split, family) %>%
spread(family, n)


This is consistent with what we have

eval_long_df %>%
drop_na(value) %>%
filter(
    WBtest_is_significant,
    model %in% c('full'),
    split %in% c('non_british_white', 'african', 's_asian', 'e_asian'),
    metric %in% c('r2', 'NagelkerkeR2')
) %>%
count(split, family) %>%
spread(family, n)


### median of the diff(incremental predictive performance in other ancestry group vs WB)

transferability_df %>%
filter(
    WBtest_is_significant,
    model == "delta"
) %>%
group_by(family, metric, split) %>% 
summarise(median = median(delta_vs_WBtest), `.groups` = 'drop') -> transferability_median_df


# add non-biomarker traits only analysis
transferability_median_df %>%
bind_rows(
    transferability_df %>%
    filter(trait_category != "Biomarkers") %>%
    filter(
        WBtest_is_significant,
        model == "delta",
        metric == "r2"
    ) %>%
    mutate(metric = "r2noBiomarkers") %>%
    group_by(family, metric, split) %>% 
    summarise(median = median(delta_vs_WBtest), `.groups` = 'drop')   
) -> transferability_median_df


transferability_median_df


transferability_median_df %>%
pivot_wider(
    names_from = c("family", "metric"),
    values_from = c("median")
) %>%
left_join(
    get_split_order_df(),
    by = "split"
) %>%
arrange(split_order)

### get the slope of regression

lapply(unique(transferability_df$metric), function(metric_selected){
    c('non_british_white', 'african', 's_asian', 'e_asian') %>%
    lapply(function(pop){
        glm(
            (delta_WBtest + delta_vs_WBtest) ~ 0 + delta_WBtest,
            data = transferability_df %>%
            filter(
                WBtest_is_significant,
                model == "delta",
                metric == metric_selected,
                split == pop
            )
        ) %>%
        fit_to_df() %>%
        mutate(split = pop, metric = metric_selected)  
    }) %>% bind_rows()
}) %>% bind_rows() %>%
bind_rows(
    # non-biomarker traits-only analysis
    c('non_british_white', 'african', 's_asian', 'e_asian') %>%
    lapply(function(pop){
        glm(
            (delta_WBtest + delta_vs_WBtest) ~ 0 + delta_WBtest,
            data = transferability_df %>%
            filter(trait_category != "Biomarkers") %>%
            filter(
                WBtest_is_significant,
                model == "delta",
                metric == "r2",
                split == pop
            )
        ) %>%
        fit_to_df() %>%
        mutate(split = pop, metric = "r2noBiomarkers")  
    }) %>% bind_rows()
) %>%
left_join(
    transferability_df %>%
    filter(
        WBtest_is_significant,
        model == "delta"
    ) %>%
    count(metric, family, split),
    by = c("split", "metric")
) %>%
left_join(
    get_split_order_df(),
    by = "split"
) %>%
arrange(family, metric, split_order) -> regression_slopes_df


regression_slopes_df %>%
select(all_of(c(
    "family", "metric", "split", "n", "variable", "estimate", "SE", "z_or_t_value", "P"
)))


## Violin plots

plot_transferability_violin <- function(transferability_df, metric_selected, ypos = c(.10, .10, .10, .10)){
    transferability_df %>%
    filter(
        WBtest_is_significant,
        model == "delta",
        metric == metric_selected
    ) %>%
    ggplot(aes(
        x=reorder(split_plot, split_order),
        y=delta_vs_WBtest,
        color=reorder(split_plot, split_order)
    )) + 
    geom_hline(yintercept = 0, color='gray') +
    geom_violin() +
    geom_jitter(height = 0, width = 0.3) +
    theme_bw(base_size = 16) + 
    scale_color_manual(values=get_split_color_dict()) +
    labs(
        x = 'Ancestry group in UK Biobank'
    ) +
    theme(legend.position = 'none') +
    stat_summary(
        fun = "median",
        geom = "crossbar", 
        width = 0.8,
        color = "black"
    ) + 
    annotate(
        geom="text", x = 1, y = ypos[1], color="black", size = 6,
        label = transferability_median_df %>%
            filter(
                split == 'non_british_white',
                metric == metric_selected
            ) %>% pull(median) %>% round(5)
    ) + 
    annotate(
        geom="text", x = 2, y = ypos[2], color="black", size = 6,
        label = transferability_median_df %>%
            filter(
                split == 's_asian',
                metric == metric_selected
            ) %>% pull(median) %>% round(5)
    ) + 
    annotate(
        geom="text", x = 3, y = ypos[3], color="black", size = 6,
        label = transferability_median_df %>%
            filter(
                split == 'e_asian',
                metric == metric_selected
            ) %>% pull(median) %>% round(5)
    ) + 
    annotate(
        geom="text", x = 4, y = ypos[4], color="black", size = 6,
        label = transferability_median_df %>%
            filter(
                split == 'african',
                metric == metric_selected
            ) %>% pull(median) %>% round(5)
    )
}



p_trans_violin_gaussian <- transferability_df %>%
filter(
    family == "gaussian"
) %>%
plot_transferability_violin("r2") +
labs(
    title = 'Quantitative traits (Gaussian model)',
    y = latex2exp::TeX("The difference in incremental \\textit{R}$^2$ compared against white British")
)

p_trans_violin_gaussian_noBiomarkers <- transferability_df %>%
filter(
    family == "gaussian"
) %>%
filter(trait_category != "Biomarkers") %>%
mutate(
    metric = if_else(metric == "r2", "r2noBiomarkers", metric)
) %>%
plot_transferability_violin("r2noBiomarkers") +
labs(
    title = 'Quantitative traits (Gaussian model)\n(non-Biomarker traits only)',
    y = latex2exp::TeX("The difference in incremental \\textit{R}$^2$ compared against white British")
)


p_trans_violin_NagelkerkeR2 <- transferability_df %>%
filter(
    family == "binomial"
) %>%
plot_transferability_violin("NagelkerkeR2") +
labs(
    title = 'Binary traits (Binomial model)',
    y = latex2exp::TeX("The difference in incremental Nagelkerke's \\textit{R}$^2$ compared against white British")
)

p_trans_violin_AUC <- transferability_df %>%
filter(
    family == "binomial"
) %>%
plot_transferability_violin("auc", c(.10, .05, .05, .05)) +
labs(
    title = 'Binary traits (Binomial model)',
    y = latex2exp::TeX("The difference in incremental AUC compared against white British")
)

p_trans_violin_TjurR2 <- transferability_df %>%
filter(
    family == "binomial"
) %>%
plot_transferability_violin("TjurR2", c(.05, .05, .05, .05)) +
labs(
    title = 'Binary traits (Binomial model)',
    y = latex2exp::TeX("The difference in incremental Tjur's \\textit{R}$^2$ compared against white British")
)

## Scatter plots

plot_transferability_scatter <- function(
    transferability_df, metric_selected,
    xpos_base = 0.5,
    ypos = c(.0, .0, .0, .0)
){
    transferability_df %>%
    filter(
        WBtest_is_significant,
        model == "delta",
        metric == metric_selected
    ) %>%
    ggplot(aes(
        x=delta_WBtest,
        y=delta_vs_WBtest + delta_WBtest,
        color=reorder(split_plot, split_order)
    )) -> plt_obj

    for(split_plot_str in names(get_split_color_dict())){
        this_slope <- regression_slopes_df %>%
        filter(
            metric == metric_selected,
            split_plot == split_plot_str
        ) %>%
        pull(estimate)
        plt_obj <- plt_obj + geom_abline(
            color=get_split_color_dict()[[split_plot_str]],
            linetype = "dashed",
            intercept=0,
            slope = this_slope
        ) +
        annotate(
            geom="text",
            x = xpos_base,
            y = xpos_base * this_slope + ypos[which(names(get_split_color_dict()) == split_plot_str)],
            color="black",
            size = 6,
            label = sprintf('y = %.03fx', this_slope)
        )
    }

    plt_obj +
    geom_point() +
    theme_bw(base_size = 16) + 
    scale_color_manual(values=get_split_color_dict()) +
    labs(
        color = 'Ancestry group in UK Biobank'
    ) +
    theme(legend.position = c(.28,.85))
}


p_trans_scatter_gaussian <- transferability_df %>%
filter(
    family == "gaussian"
) %>%
plot_transferability_scatter(
    "r2",
    xpos_base = .3,
    ypos = c(0, 0.01, -0.01, 0)
) +
labs(
    title = 'Quantitative traits (Gaussian model)',
    x = latex2exp::TeX('Incremental \\textit{R}$^2$ in white British'),
    y = latex2exp::TeX('Incremental \\textit{R}$^2$')
)


p_trans_scatter_gaussian_noBiomarkers <- transferability_df %>%
filter(
    family == "gaussian"
) %>%
filter(trait_category != "Biomarkers") %>%
mutate(
    metric = if_else(metric == "r2", "r2noBiomarkers", metric)
) %>%
plot_transferability_scatter(
    "r2noBiomarkers",
    xpos_base = .3,
    ypos = c(0, 0.01, -0.01, 0)
) +
labs(
    title = 'Quantitative traits (Gaussian model)\n(non-Biomarker traits only)',
    x = latex2exp::TeX('Incremental \\textit{R}$^2$ in white British'),
    y = latex2exp::TeX('Incremental \\textit{R}$^2$')
)


p_trans_scatter_NagelkerkeR2 <- transferability_df %>%
filter(
    family == "binomial"
) %>%
plot_transferability_scatter(
    "NagelkerkeR2",
    ypos = c(0, -0.03, 0.02, 0)
) +
labs(
    title = 'Binary traits (Binomial model)',
    x = latex2exp::TeX("Incremental Nagelkerke's \\textit{R}$^2$ in white British"),
    y = latex2exp::TeX("Incremental Nagelkerke's \\textit{R}$^2$")
)

p_trans_scatter_TjurR2 <- transferability_df %>%
filter(
    family == "binomial"
) %>%
plot_transferability_scatter(
    "TjurR2",
    xpos_base = .35,
    ypos = c(0, 0.025, -0.03, 0.0)
) +
labs(
    title = 'Binary traits (Binomial model)',
    x = latex2exp::TeX("Incremental Tjur's \\textit{R}$^2$ in white British"),
    y = latex2exp::TeX("Incremental Tjur's \\textit{R}$^2$")
)

p_trans_scatter_AUC <- transferability_df %>%
filter(
    family == "binomial"
) %>%
plot_transferability_scatter(
    "auc",
    xpos_base = .35,
    ypos = c(0, 0.01, 0, 0)
) +
labs(
    title = 'Binary traits (Binomial model)',
    x = latex2exp::TeX("Incremental AUC in white British"),
    y = latex2exp::TeX("Incremental AUC")
)

for(ext in c('png', 'pdf')){
    ggsave(
        file.path("plots", sprintf('transferability.%s', ext)),
        gridExtra::arrangeGrob(
            p_trans_violin_NagelkerkeR2 +
            labs(title = '(A) Binary traits (Binomial model)'),
            p_trans_scatter_NagelkerkeR2 +
            labs(title = '(B) Binary traits (Binomial model)'),
            p_trans_violin_gaussian +
            labs(title = '(C) Quantitative traits (Gaussian model)'),
            p_trans_scatter_gaussian +
            labs(title = '(D) Quantitative traits (Gaussian model)'),
            ncol=2
        ),
        width=16, height=16
    )
    ggsave(
        file.path("plots", sprintf('transferability_gaussian_noBiomarkers.%s', ext)),
        gridExtra::arrangeGrob(
            p_trans_violin_gaussian_noBiomarkers,
            p_trans_scatter_gaussian_noBiomarkers,
            ncol=2
        ),
        width=16, height=8
    )
    ggsave(
        file.path("plots", sprintf('transferability_AUC.%s', ext)),
        gridExtra::arrangeGrob(
            p_trans_violin_AUC,
            p_trans_scatter_AUC,
            ncol=2
        ),
        width=16, height=8
    )
    ggsave(
        file.path("plots", sprintf('transferability_Tjur.%s', ext)),
        gridExtra::arrangeGrob(
            p_trans_violin_TjurR2,
            p_trans_scatter_TjurR2,
            ncol=2
        ),
        width=16, height=8
    )
}
