==> src/20211028_freeze/8_population_characteristics.ipynb <==
suppressWarnings(suppressPackageStartupMessages({
    library(tidyverse)
    library(data.table)
}))


source('paths.sh')

PRS202110_f %>%
fread(
    colClasses = c('#FID'='character', 'IID'='character'),
    select=c('#FID', 'IID', 'population', 'split', 'age', 'age0', 'sex', 'Array')
) %>%
rename_with(
    function(x){str_replace(x, '#', '')}, starts_with("#")
) -> phe_all_df


pops <- c('white_british', 'non_british_white', 'african', 's_asian', 'e_asian')


phe_all_df %>%
filter(population %in% pops) -> phe_df


phe_df %>%
nrow

phe_df %>% 
count(population, split) %>% arrange(-n)

235991 + 33713

235991 + 33713 + 67425

phe_df %>%
group_by(split) %>%
summarise(
    n = n(),
    age0_mean = mean(age0),
    age0_sd = sd(age0),
    sex_M = mean(sex),
    Array_UKBB = mean(Array)
) %>%
left_join(
    phe_df %>% 
    count(population, split), 
    by = c("split", "n")
) %>%
arrange(-n) %>%
bind_rows(
    phe_df %>%
    summarise(
        n = n(),
        age0_mean = mean(age0),
        age0_sd = sd(age0),
        sex_M = mean(sex),
        Array_UKBB = mean(Array)
    ) %>%
    mutate(population = 'Total', split='N/A')
) %>%
select(
    population, split, n, age0_mean, age0_sd, sex_M, Array_UKBB
) -> pops_tbl_df


pops_tbl_df %>%
rename('#population' = 'population') %>%
fwrite('pop_characteristics.tsv', sep='\t', na = "NA", quote=F)
